dev: 1.初步完成H5/小程序用户端的用户体系介入；

+++ farm_backend/farms-system/src/main/java/com/neo/system/service/impl/SysRoleServiceImpl.java
@@ -123,7 +123,7 @@ public class SysRoleServiceImpl implements ISysRoleService
      * @return 选中角色ID列表
      */
     @Override
-    public List<Long> selectRoleListByUserId(Long userId)
+    public List<Integer> selectRoleListByUserId(Long userId)
     {
         return roleMapper.selectRoleListByUserId(userId);
     }
